// client/src/app/model/template-contagem.ts
export class TemplateContagem {
    
    id: number;
    projetoAba: string;
    projetoReferencia: string;
    linguagemAba: string;
    linguagemReferencia: string;
    tipoProjetoAba: string;
    tipoProjetoReferencia: string;
    tipoContagemAba: string;
    tipoContagemReferencia: string;
    plataformaAba: string;
    plataformaReferencia: string;
    liderProjetoAba: string;
    liderProjetoReferencia: string;
    observacoesAba: string;
    observacoesReferencia: string;
    artefatosUsadosAba: string;
    artefatorUsadosReferencia: string;
    propositoEscopoAba: string;
    propositoEscopoReferencia: string;
    fronteiraAba: string;
    fronteiraReferencia: string;
    totalPfDemandaAba: string;
    totalPfDemandaReferencia: string;
    pfFuncaoDadosAba: string;
    pfFuncaoDadosReferencia: string;
    pfRetrabalhoAba: string;
    pfRetrabalhoReferencia: string;
    pfFuncaoTransacionalAba: string;
    pfFuncaoTransacionalReferencia: string;
    totalPfBrutoAba: string;
    totalPfBrutoReferencia: string;
    scopeCreepAba: string;
    scopeCreepReferencia: string;
    metodoContagemAba: string;
    metodoContagemReferencia: string;

    constructor() {

    }
}